move ToPlural to rules/internal/utils/string_pluralize.go
