+++ rules/aip0158/response_plural_first_field.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/googleapis/api-linter/lint"
 	"github.com/googleapis/api-linter/locations"
+	"github.com/googleapis/api-linter/rules/internal/utils"
 	"github.com/jhump/protoreflect/desc"
 )
 
@@ -28,7 +29,7 @@ var responsePluralFirstField = &lint.MessageRule{
 	LintMessage: func(m *desc.MessageDescriptor) []lint.Problem {
 		// Throw a linter warning if, the first field in the message is not named according to plural(message_name.to_snake().split('_')[1:-1]).
 		firstField := m.GetFields()[0]
-		want := lint.ToPlural(firstField.GetName())
+		want := utils.ToPlural(firstField.GetName())
 		if want != firstField.GetName() {
 			return []lint.Problem{{
 				Message:    fmt.Sprintf("First field of Paginated RPCs' response should be plural."),
